requirements.txt
arxiv>=2.1.0
requests>=2.31.0
openai>=1.30.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
groq>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0